Move slash functions to a separate file
