+++ src/opt/single.c
@@ -16,6 +16,7 @@
 #include "../builtins/fns.c"
 #include "../builtins/sfns.c"
 #include "../builtins/select.c"
+#include "../builtins/slash.c"
 #include "../builtins/sysfn.c"
 #include "../builtins/sort.c"
 #include "../builtins/selfsearch.c"
